+++ src/lib.rs
@@ -48,7 +48,7 @@
 //! ```rust
 //! use libmcaptcha::{
 //!     cache::{messages::VerifyCaptchaResult, HashCache},
-//!     master::{AddSiteBuilder, Master},
+//!     master::embedded::{AddSiteBuilder, Master},
 //!     pow::{ConfigBuilder, Work},
 //!     system::SystemBuilder,
 //!     DefenseBuilder, LevelBuilder, MCaptchaBuilder,