+++ CHANGELOG.md
@@ -8,6 +8,9 @@
   internet. Ideally `mCaptcha` should automatically generate random
   salt and rotate periodically, maybe in the next version.
 
+- `master::Master` is moved to `master::embedded::Master` in preparation
+  for Redis based implementation.
+
 ## 0.1.3
 
 ## Added